[refactor] 타입 변경

+++ src/main/java/com/rence/office/model/OfficeInfoVO.java
@@ -81,5 +81,5 @@ public class OfficeInfoVO {
 	private String detail_address;
 	
 	@Column(name="avg_rating")
-	private Float avg_rating;	
+	private Double avg_rating;	
 }
